requirements.txt
Package                Version  
---------------------- ---------
bcrypt                 3.1.7    
certifi                2019.6.16
cffi                   1.12.3   
chardet                3.0.4    
Django                 2.2.3    
django-cors-headers    3.0.2    
idna                   2.8      
mysql-connector-python 8.0.16   
pip                    19.1.1   
protobuf               3.9.0    
pycparser              2.19     
PyJWT                  1.7.1    
pytz                   2019.1   
requests               2.22.0   
setuptools             41.0.1   
six                    1.12.0   
sqlparse               0.3.0    
urllib3                1.25.3   
wheel                  0.33.4   
